pyproject.toml
[project]
name = "unitn-sde"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python-ng",
    "opencv-python",
    "psutil",
    "python-ffmpeg",
    "requests",
]

[tool.pyright]
pythonVersion = "3.12"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN1",   # Missing type self, cls (deprecated)
    "ANN204", # Missing type annotation for special method
    "D1",     # Missing docstring in public: module, class, method, function, package, magic method, nested class, init
    "D203",   # 1 blank line required before class docstring
    "D212",   # Multi-line docstring summary should start at the first line
    "DTZ",    # No Naive Datetime
    "E501",   # Line too long
    "S311",   # Standard pseudo-random generators are not suitable for cryptographic purposes
    "T201",   # Checks for print statements.
    "TRY400", # Use logging.exception() instead of logging.error() for logging exceptions
]
